+++ eslint.config.js
@@ -47,6 +47,7 @@ export default defineConfig([
     rules: {
       'no-unused-expressions': 'off',
       'import/no-unresolved': ['error', { ignore: ['@octokit/rest', 'is-language-code', 'uuid'] }],
+      curly: 'off',
     },
   },
   {